backend/requirements.txt
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.0
pydantic-settings>=2.0
requests>=2.31
redis>=5.0
celery>=5.3

# Utilities
pyyaml>=6.0
tqdm>=4.66
openai>=1.0
python-dotenv>=1.0
replicate>=0.20.0

# Image/Video Processing
pillow>=10.0
numpy>=1.24
opencv-python>=4.8
av>=11.0.0
scikit-image>=0.22.0
ultralytics>=8.0.0
numpy>=1.24
opencv-python>=4.8
av>=11.0.0
scikit-image>=0.22.0

# AI/ML (GPU)
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0
diffusers>=0.25.0
transformers>=4.36.0
accelerate>=0.25.0
optimum>=1.16.0
bitsandbytes>=0.41.0
huggingface-hub>=0.20.0

# Testing
pytest>=7.0.0
httpx>=0.23.0

# Agentic (LangGraph)
langgraph>=0.0.10
langchain>=0.1.0
langchain-openai>=0.0.5